impletmented the coding for movement audio
with the audio i imported into unity, i then coded the player.cs so that the jump sound would play on every jump. i decided that it would be best if i kept the pitch the same. i also decided to remove the log jump audio as i thought it might complicate things later on